Add Output Logging Module
Add the BuildServe.Logging module, which contains types to describe the
output queue, & log messages, functions to log messages, and functions
to print messages to stdout.

There is a generic logging function, as well a function for logging the
output of handler(tagged with a LogType) and specialized versions for
both Client & Server command handles.