Add BuildServe.Runner Module
Add the Runner module, which contains functions to assist in running
commands with output logging that utilizes the Logging module.

The run function lets us run arbitrary commands in a specific working
directory, logging the output with a specific tag. installDependency is
a specialized version of run for installing the dependencies of the
Client or Server. removeFiles is useful for writing clean-up commands.