Refactor the Logging Typeclass
Modify the HasLogQueue typeclass so that getLogQueue simply pulls a
LogQueue out of some type.

Add a new MonadLoggable typeclass for actually logging messages, move
the logMessage function into the typeclass(as a method), and provide a
default implementation for `ReaderT env m` where `env` has a
`HasLogQueue` instance.

Fix the type declarations for functions that used the HasLogQueue
constraint, using the new MonadLoggable constraint instead.